requirements.txt
chainlit==2.9.6
langchain-openai==1.1.9
langchain-community==0.4.1
langgraph==1.0.8
faiss-cpu==1.13.2
psycopg2-binary==2.9.11
python-dotenv==1.2.1
openai==1.109.1
fastapi==0.128.8
uvicorn==0.40.0
pydantic==2.9.2
pypdf==6.6.2